fix: 视频抽帧 fps 整型化修「端点 400」；失败日志附端点响应体；skill 导入补路径与单文件入口
- 视频 fps 一直以浮点（2.0）进请求体，SiliconFlow 校验判 20015 非法、视频打标必 400——
  内部类型链 float→int（VideoPart / engine / CLI / 前端控件），HTTP 契约保持 number 并在
  LabelRequest 校验器兜底整性（非整数 422），避免 oasdiff ERR 级破坏性变更挡板
- llm 失败日志（complete / stream）附端点响应体摘要（截 300 字符），「模型侧 400」不再盲猜
- /api/skills/import 扩展：路径指向单个 .md 文件时按 SKILL.md 单文件导入；
  技能页补「SKILL.md 文件…」「粘贴路径 → 导入」两个入口（保留文件夹上传）

门禁：后端 380 过 4 跳 / pyright 0 / ruff 0 / mkdocs 过 / 契约快照重导出（oasdiff 本地预检无
破坏性变更）；前端 lint + typecheck + 38 测试 + dist 重建；e2e 4 项过。

+++ backend/src/dataset_factory/labeling/engine.py
@@ -148,7 +148,7 @@ class LabelingEngine:
         video_bytes: bytes | None = None,
         video_name: str = "video.mp4",
         video_mime: str = "video/mp4",
-        video_fps: float = 2.0,
+        video_fps: int = 2,
         video_max_frames: int = 16,
     ) -> LabelResult:
         """跑一轮打标：组装请求 → 先落信封 → 调模型 → 落回复，返回 caption 与会话 id。
@@ -259,7 +259,7 @@ class LabelingEngine:
         video_bytes: bytes | None = None,
         video_name: str = "video.mp4",
         video_mime: str = "video/mp4",
-        video_fps: float = 2.0,
+        video_fps: int = 2,
         video_max_frames: int = 16,
     ) -> Iterator[StreamStarted | StreamDelta | StreamFinished]:
         """流式跑一轮打标：先落信封 → 逐段产出增量 → 终稿落盘，事件序列返回给调用方。
@@ -569,7 +569,7 @@ def _assemble(
     image_bytes: bytes | None,
     video_bytes: bytes | None,
     video_mime: str,
-    video_fps: float,
+    video_fps: int,
     video_max_frames: int,
     attachment: str | None,
 ) -> tuple[list[Message], list[JsonValue]]:
